% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{comparimotif}
\alias{comparimotif}
\title{Simple motif comparison}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
comparimotif(motifs = NULL, cmotifs = NULL)
}
\arguments{
\item{motifs}{A \code{regex} representing motifs.}

\item{cmotifs}{A \code{regex} representing motifs to compare to
\code{motifs}.}
}
\value{
A \code{tibble}.
}
\description{
Simple motif comparison
}
