% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{gopher}
\alias{gopher}
\title{Generate alignment for orthologous proteins}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
gopher(uniprotid = NULL, taxonid = 9606)
}
\arguments{
\item{uniprotid}{A \code{character} vector of one or more UniProt ID.}

\item{taxonid}{A \code{numeric}. Default is 9606 NCBI species ID.}
}
\value{
A \code{tibble}.
}
\description{
Generate alignment for orthologous proteins
}
