% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{peploc}
\alias{peploc}
\title{Identify bioactive peptides in protein sequences}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
peploc(uniprotid = NULL)
}
\arguments{
\item{uniprotid}{A \code{character} vector of one or more UniProt ID.}
}
\value{
A \code{tibble}.
}
\description{
Identify bioactive peptides in protein sequences
}
