% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biowareapi.R
\docType{data}
\name{query_map_bioware}
\alias{query_map_bioware}
\title{Query Map for bioware API}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13 rows and 7 columns.}
\usage{
query_map_bioware
}
\description{
The object documents the different types of requests and their query
paramaters that are allowed by the API.
}
\keyword{datasets}
