% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{slimfinder}
\alias{slimfinder}
\title{Short Linear Motif Finder}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
slimfinder(uniprotid = NULL, dismask = TRUE)
}
\arguments{
\item{uniprotid}{A \code{character} vector of one or more UniProt ID.}

\item{dismask}{A \code{logical} default (TRUE) of disorder masking.}
}
\value{
A \code{tibble}.
}
\description{
Short Linear Motif Finder
}
\seealso{
slimprints

slimsearch

slimpred
}
