% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{slimpred}
\alias{slimpred}
\title{Identify potential SLiM-like region in a protein sequence}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
slimpred(uniprotid = NULL)
}
\arguments{
\item{uniprotid}{A \code{character} vector of one or more UniProt ID.}
}
\value{
A \code{tibble}.
}
\description{
Identify potential SLiM-like region in a protein sequence
}
\seealso{
slimprints

slimfinder

slimsearch
}
