% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{slimprints}
\alias{slimprints}
\title{Find cohorts of locally conserved residues}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
slimprints(uniprotid = NULL)
}
\arguments{
\item{uniprotid}{A \code{character} vector of one or more UniProt ID.}
}
\value{
A \code{tibble}.
}
\description{
Find cohorts of locally conserved residues
}
\seealso{
slimfinder

slimsearch

slimpred
}
