% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{slimsearch}
\alias{slimsearch}
\title{Proteome side motif search}
\source{
http://bioware.ucd.ie/~compass/biowareweb/
}
\usage{
slimsearch(uniprotid = NULL, motif, taxonid = 9606,
  orthdb = "Orthdb")
}
\arguments{
\item{uniprotid}{A \code{character} vector of one or more UniProt ID.}

\item{motif}{A \code{regex} for a motif to search.}

\item{taxonid}{A \code{numeric}. Default is 9606 NCBI species ID.}

\item{orthdb}{A \code{character} only 'Orthdb'}
}
\value{
A \code{tibble}.
}
\description{
Proteome side motif search
}
\seealso{
slimprints

slimfinder

slimpred
}
